requirements.txt
Flask
tensorflow
numpy
Pillow
gunicorn  # Essential for production deployment